Layout task network diagram
Attempts to visualise the task network in the 'Task Network' tab.
Hierarchies & dependencies are both shown, with tasks represented as
cylinders. Hovering over a cylinder will reveal the task name, and
clicking on it will cause the selected task to appear in the context
pane. Heuristic approach used to find a network layout that minimises
edge crossover and distance.

Really quite proud of this.

+++ graft/layers/presentation/tkinter_gui/task_network_graph_drawing/depth_position_assignment/position.py
@@ -9,7 +9,7 @@ class RelationPosition:
     def __init__(
         self,
         dependency_position: dependency_position.DependencyPosition,
-        hierarchy_position: float,
+        hierarchy_position: int,
     ) -> None:
         self._dependency = dependency_position
         self._hierarchy = hierarchy_position
@@ -19,5 +19,5 @@ class RelationPosition:
         return self._dependency
 
     @property
-    def hierarchy(self) -> float:
+    def hierarchy(self) -> int:
         return self._hierarchy
